requirements.txt
fastapi
uvicorn
langchain
langchain_community
langgraph
langchain_openai
tavily-python
openai
requests
pydantic
python-dotenv
wikipedia





